arbitrary steps
the slider can now take an ordered array of numbers to use as values. it
can also use an ordered array of objects that have a `value` property,
which is used for the value.

when using the `steps` attribute, the `min`, `max`, and `step` attributes
are ignored. instead, the first and last items in the `steps` array are
used for the `min` and `max`, and a step is not incremental. instead, a
step corresponds to a value in the `steps` array.